refactor: 뷰홀더 분리
