+++ android/app/src/main/java/com/teameetmeet/meetmeet/presentation/calendar/monthcalendar/EventBarAdapter.kt
@@ -1,46 +1,17 @@
 package com.teameetmeet.meetmeet.presentation.calendar.monthcalendar
 
-import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.annotation.Dimension
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import com.teameetmeet.meetmeet.databinding.ItemEventBarBinding
 import com.teameetmeet.meetmeet.presentation.model.EventBar
 
-class EventBarAdapter : ListAdapter<EventBar?, EventBarAdapter.EventBarViewHolder>(diffCallback) {
-    override fun onCreateViewHolder(
-        parent: ViewGroup,
-        viewType: Int
-    ): EventBarViewHolder {
-        val binding = ItemEventBarBinding.inflate(
-            LayoutInflater.from(parent.context),
-            parent,
-            false
-        )
-        binding.root.layoutParams.height = (parent.measuredHeight / 5)
-        binding.itemEventBarTvHidden.setTextSize(
-            Dimension.PX, parent.measuredHeight.toFloat() / 20
-        )
-        return EventBarViewHolder(binding)
+class EventBarAdapter : ListAdapter<EventBar?, EventBarViewHolder>(diffCallback) {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): EventBarViewHolder {
+        return EventBarViewHolder.from(parent)
     }
 
     override fun onBindViewHolder(holder: EventBarViewHolder, position: Int) {
         holder.bind(getItem(position))
-    }
-
-    class EventBarViewHolder(
-        private val binding: ItemEventBarBinding,
-    ) : RecyclerView.ViewHolder(binding.root) {
-        fun bind(item: EventBar?) {
-            binding.item = item
-            binding.itemEventBarTvHidden.text =
-                with(item) {
-                    if (this != null && hiddenCount > 0) "+$hiddenCount"
-                    else ""
-                }
-        }
     }
 
     companion object {
