+++ android/app/src/main/java/com/teameetmeet/meetmeet/presentation/calendar/monthcalendar/MonthCalendarAdapter.kt
@@ -1,47 +1,20 @@
 package com.teameetmeet.meetmeet.presentation.calendar.monthcalendar
 
-import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import com.teameetmeet.meetmeet.databinding.ItemCalendarBinding
 import com.teameetmeet.meetmeet.presentation.model.CalendarItem
 
 class MonthCalendarAdapter(
-    private val calendarItemClickListener: CalendarItemClickListener
-) : ListAdapter<CalendarItem, MonthCalendarAdapter.CalendarViewHolder>(diffCallback) {
+    private val listener: CalendarItemClickListener
+) : ListAdapter<CalendarItem, MonthCalendarViewHolder>(diffCallback) {
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CalendarViewHolder {
-        val binding = ItemCalendarBinding.inflate(
-            LayoutInflater.from(parent.context),
-            parent,
-            false
-        )
-        binding.root.layoutParams.height = parent.measuredHeight / 6
-        return CalendarViewHolder(binding, calendarItemClickListener)
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MonthCalendarViewHolder {
+        return MonthCalendarViewHolder.from(parent, listener)
     }
 
-    override fun onBindViewHolder(holder: CalendarViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: MonthCalendarViewHolder, position: Int) {
         holder.bind(getItem(position))
-    }
-
-    class CalendarViewHolder(
-        private val binding: ItemCalendarBinding,
-        private val calendarItemClickListener: CalendarItemClickListener
-    ) : RecyclerView.ViewHolder(binding.root) {
-        fun bind(item: CalendarItem) {
-            with(binding) {
-                this.item = item
-                itemCalendarViewTouch.setOnClickListener {
-                    calendarItemClickListener.onItemClick(item)
-                }
-                if (itemCalendarRvEvents.adapter == null) {
-                    itemCalendarRvEvents.adapter = EventBarAdapter()
-                }
-                itemCalendarRvEvents.itemAnimator = null
-            }
-        }
     }
 
     companion object {
