+++ android/app/src/main/java/com/teameetmeet/meetmeet/presentation/eventstory/createfeed/MediaViewHolder.kt
@@ -1,5 +1,7 @@
 package com.teameetmeet.meetmeet.presentation.eventstory.createfeed
 
+import android.view.LayoutInflater
+import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.teameetmeet.meetmeet.databinding.ItemCreateFeedMediaBinding
 import com.teameetmeet.meetmeet.presentation.model.FeedMedia
@@ -12,6 +14,16 @@ class MediaViewHolder(
         binding.mediaItem = item
         binding.itemCreateFeedIbCancel.setOnClickListener {
             listener.onItemClick(item)
+        }
+    }
+
+    companion object {
+        fun from(parent: ViewGroup, listener: MediaItemCancelClickListener): MediaViewHolder {
+            val inflater = LayoutInflater.from(parent.context)
+            return MediaViewHolder(
+                ItemCreateFeedMediaBinding.inflate(inflater, parent, false),
+                listener
+            )
         }
     }
 }
