--- android/app/src/main/java/com/teameetmeet/meetmeet/presentation/eventstory/feeddetail/FeedCommentsAdapter.kt
@@ -1,59 +1,0 @@
-package com.teameetmeet.meetmeet.presentation.eventstory.feeddetail
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.core.view.isVisible
-import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import com.teameetmeet.meetmeet.data.model.Comment
-import com.teameetmeet.meetmeet.databinding.ItemFeedCommentBinding
-import com.teameetmeet.meetmeet.presentation.model.EventAuthority
-
-class FeedCommentsAdapter(
-    private val authority: EventAuthority,
-    private val commentDeleteClickListener: CommentDeleteClickListener
-) :
-    ListAdapter<Comment, FeedCommentsAdapter.FeedCommentsViewHolder>(diffCallback) {
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): FeedCommentsViewHolder {
-        val binding = ItemFeedCommentBinding.inflate(
-            LayoutInflater.from(parent.context),
-            parent,
-            false
-        )
-        return FeedCommentsViewHolder(binding, authority, commentDeleteClickListener)
-    }
-
-    override fun onBindViewHolder(holder: FeedCommentsViewHolder, position: Int) {
-        holder.bind(getItem(position))
-    }
-
-    class FeedCommentsViewHolder(
-        private val binding: ItemFeedCommentBinding,
-        private val authority: EventAuthority,
-        private val commentDeleteClickListener: CommentDeleteClickListener
-    ) :
-        RecyclerView.ViewHolder(binding.root) {
-        fun bind(item: Comment) {
-            binding.item = item
-            binding.feedCommentBtnDelete.isVisible =
-                item.isMine || authority == EventAuthority.OWNER
-            binding.feedCommentBtnDelete.setOnClickListener {
-                commentDeleteClickListener.onClick(item)
-            }
-        }
-    }
-
-    companion object {
-        private val diffCallback = object : DiffUtil.ItemCallback<Comment>() {
-            override fun areItemsTheSame(oldItem: Comment, newItem: Comment): Boolean {
-                return oldItem.id == newItem.id
-            }
-
-            override fun areContentsTheSame(oldItem: Comment, newItem: Comment): Boolean {
-                return oldItem == newItem
-            }
-
-        }
-    }
-}
